requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
boto3==1.29.7
python-decouple==3.8
Pillow==10.1.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
python-multipart==0.0.6
dropbox==11.36.2
pydantic==2.5.0
celery==5.3.4
redis==5.0.1
django-storages==1.14.2
psycopg2-binary==2.9.9